move: [order] kafka 관련 파일 폴더구조 변경
