+++ order/src/main/java/com/high/order/infrastructure/kafka/consumer/KafkaConsumer.java
@@ -3,13 +3,13 @@ package com.high.order.infrastructure.kafka.consumer;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.high.order.application.OrderServiceV2;
-import com.high.order.application.dto.internal.kafka.request.CreateOrderCommand;
-import com.high.order.application.dto.internal.kafka.request.DeleteOrderCommand;
-import com.high.order.application.dto.internal.kafka.response.OrderCreateFailedResponse;
-import com.high.order.application.dto.internal.kafka.response.OrderDeleteResponse;
-import com.high.order.application.dto.internal.kafka.response.OrderSuccessResponse;
+import com.high.order.application.dto.event.request.CreateOrderCommand;
+import com.high.order.application.dto.event.request.DeleteOrderCommand;
+import com.high.order.application.dto.event.response.OrderCreateFailedResponse;
+import com.high.order.application.dto.event.response.OrderDeleteResponse;
+import com.high.order.application.dto.event.response.OrderSuccessResponse;
 import com.high.order.application.port.EventPublisher;
-import com.high.order.infrastructure.adapter.OrderCreateAdapter;
+import com.high.order.infrastructure.kafka.adapter.OrderCreateAdapter;
 import com.high.order.infrastructure.context.MessageContext;
 import com.high.order.infrastructure.exception.EmptyKafkaMessageException;
 import com.high.order.infrastructure.kafka.dto.response.OrderCreateRequestMessage;
